+++ core/TreeNode.ts
@@ -2828,7 +2828,7 @@ class TreeNode extends AbstractNode {
     return str ? indent + str.replace(/\n/g, indent) : ""
   }
 
-  static getVersion = () => "47.0.0"
+  static getVersion = () => "47.1.0"
 
   static fromDisk(path: string): TreeNode {
     const format = this._getFileFormat(path)